turisago-bi-service/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database connections
psycopg2-binary==2.9.9
pymysql==1.1.0
sqlalchemy==2.0.23

# Data analysis and visualization
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# Machine Learning
scikit-learn==1.3.2
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6

# CORS
fastapi-cors==0.0.6 